backend/requirements.txt
Django==4.2.16
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
psycopg2-binary==2.9.9
dj-database-url==2.1.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
Pillow==10.2.0
PyJWT==2.8.0
stripe==8.5.0
whitenoise==6.6.0
gunicorn==21.2.0
pytesseract==0.3.10
pdf2image==1.16.3
img2pdf==0.4.4
openai==1.51.2
httpx[http2]==0.27.2
python-magic==0.4.27
